Extracted the CalloutAlert component into a separate file to ease testing.
