+++ src/CalloutAlert.js
@@ -1,0 +1,27 @@
+import React from 'react'
+import PropTypes from 'prop-types'
+
+const CalloutAlert = ({error}) =>
+  <div className={`row`}>
+    <div className={`columns large-9 large-centered`}>
+      <div className={`callout alert small`}>
+        <h5>Oops!</h5>
+        <p>
+          {error.description}<br/>
+          If the error persists, in order to help us debug the issue, please copy the URL and this message and
+          send it to us via <a href={`https://www.ebi.ac.uk/support/gxasc`}>the EBI Support & Feedback system</a>:
+        </p>
+        <code>{`${error.name}: ${error.message}`}</code>
+      </div>
+    </div>
+  </div>
+
+CalloutAlert.propTypes = {
+  error: PropTypes.shape({
+    description: PropTypes.string.isRequired,
+    name: PropTypes.string.isRequired,
+    message: PropTypes.string.isRequired
+  })
+}
+
+export default CalloutAlert
