+++ src/HeatmapView.js
@@ -1,33 +1,12 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-import URI from 'urijs'
 import MarkerGeneHeatmap from './MarkerGeneHeatmap'
 import PlotSettingsDropdown from './PlotSettingsDropdown'
 import LoadingOverlay from './LoadingOverlay'
+import CalloutAlert from './CalloutAlert'
+
+import URI from 'urijs'
 import _ from 'lodash'
-
-const CalloutAlert = ({error}) =>
-  <div className={`row`}>
-    <div className={`columns large-9 large-centered`}>
-      <div className={`callout alert small`}>
-        <h5>Oops!</h5>
-        <p>
-          {error.description}<br/>
-          If the error persists, in order to help us debug the issue, please copy the URL and this message and
-          send it to us via <a href={`https://www.ebi.ac.uk/support/gxasc`}>the EBI Support & Feedback system</a>:
-        </p>
-        <code>{`${error.name}: ${error.message}`}</code>
-      </div>
-      </div>
-  </div>
-
-CalloutAlert.propTypes = {
-  error: PropTypes.shape({
-    description: PropTypes.string.isRequired,
-    name: PropTypes.string.isRequired,
-    message: PropTypes.string.isRequired
-  })
-}
 
 class HeatmapView extends React.Component {
   constructor(props) {
